// src/types/inventory.ts
export interface Vendor {
  id: string;
  name: string;
  category: ('Pembelian' | 'Servis' | 'Kalibrasi')[];
  picName: string;
  phone: string;
  email: string;
  address: string;
  contractNo: string;
  contractExpiry: string;
  rating: number; // 1 to 5
  notes?: string;
}